requirements.txt
# Direct dependencies only - let pip resolve transitive dependencies automatically

# Core AI SDKs
openai>=2.14.0
openai-agents>=0.1.0
anthropic>=0.45.0
instructor>=1.13.0

# Agent Frameworks
langchain>=1.0.0
langgraph>=1.0.0
crewai>=1.0.0
pydantic-ai>=1.0.0
llama-index>=0.12.0
atomic-agents>=2.0.0
google-adk>=0.1.0
smolagents>=1.0.0

# Tools
tavily-python>=0.5.0

# UI
streamlit>=1.42.0

# Utilities
python-dotenv>=1.0.0
